Test: 배송지 도메인 테스트 코드 작성 및 리팩토링

+++ src/main/java/com/samyookgoo/palgoosam/deliveryaddress/service/DeliveryAddressService.java
@@ -1,51 +1,77 @@
 package com.samyookgoo.palgoosam.deliveryaddress.service;
 
+import com.samyookgoo.palgoosam.auth.service.AuthService;
 import com.samyookgoo.palgoosam.deliveryaddress.domain.DeliveryAddress;
+import com.samyookgoo.palgoosam.deliveryaddress.dto.DeliveryAddressRequestDto;
+import com.samyookgoo.palgoosam.deliveryaddress.dto.DeliveryAddressResponseDto;
+import com.samyookgoo.palgoosam.deliveryaddress.exception.DeliveryAddressNotFoundException;
 import com.samyookgoo.palgoosam.deliveryaddress.repository.DeliveryAddressRepository;
+import com.samyookgoo.palgoosam.global.exception.ErrorCode;
 import com.samyookgoo.palgoosam.user.domain.User;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
-
+import com.samyookgoo.palgoosam.user.exception.UserForbiddenException;
+import com.samyookgoo.palgoosam.user.exception.UserUnauthorizedException;
 import java.util.List;
+import java.util.stream.Collectors;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
 public class DeliveryAddressService {
     private final DeliveryAddressRepository deliveryAddressRepository;
+    private final AuthService authService;
 
-    public List<DeliveryAddress> getDeliveryAddressByUserId(Long userId) {
-        return deliveryAddressRepository.findAllByUser_Id(userId);
+    public List<DeliveryAddressResponseDto> getUserDeliveryAddresses() {
+        User user = getAuthenticatedUser(authService.getCurrentUser());
+        List<DeliveryAddress> deliveryAddressList = deliveryAddressRepository.findAllByUser_Id(user.getId());
+
+        return deliveryAddressList.stream()
+                .map(DeliveryAddressResponseDto::of)
+                .collect(Collectors.toList());
     }
 
-    public boolean modifyDefaultDeliveryAddress(User user, Long addressId) {
-        DeliveryAddress deliveryAddress = deliveryAddressRepository
-                .findByUserAndId(user, addressId)
-                .orElseThrow(() -> new ResponseStatusException(
-                        HttpStatus.NOT_FOUND, "변경할 주소를 찾을 수 없습니다."
-                ));
+    @Transactional
+    public void deleteUserDeliveryAddress(Long deliveryAddressId) {
+        User user = getAuthenticatedUser(authService.getCurrentUser());
+        DeliveryAddress deliveryAddress = deliveryAddressRepository.findByUserAndId(user, deliveryAddressId)
+                .orElseThrow(() -> new DeliveryAddressNotFoundException(ErrorCode.DELIVERY_ADDRESS_NOT_FOUND));
 
-        deliveryAddressRepository.unsetAllDefaults(user.getId());
-        deliveryAddressRepository.updateDefault(deliveryAddress.getId(), user.getId());
-
-        return true;
+        if (deliveryAddress.hasPermission(user.getId())) {
+            deliveryAddressRepository.deleteById(deliveryAddress.getId());
+        } else {
+            throw new UserForbiddenException();
+        }
     }
 
-    public boolean saveUserDeliveryAddress(DeliveryAddress deliveryAddress) {
-        DeliveryAddress saved = deliveryAddressRepository.save(deliveryAddress);
-        return true;
+    @Transactional
+    public void postUserDeliveryAddress(DeliveryAddressRequestDto requestDto) {
+        User user = getAuthenticatedUser(authService.getCurrentUser());
+        DeliveryAddress isDeliveryAddressSaveSuccess = DeliveryAddress.from(requestDto, user);
+        deliveryAddressRepository.save(isDeliveryAddressSaveSuccess);
     }
 
-    public boolean deleteUserDeliveryAddress(User user, Long addressId) {
-        DeliveryAddress deliveryAddress = deliveryAddressRepository
-                .findByUserAndId(user, addressId)
-                .orElseThrow(() -> new ResponseStatusException(
-                        HttpStatus.NOT_FOUND, "삭제할 주소를 찾을 수 없습니다."
-                ));
+    @Transactional
+    public void modifyDefaultDeliveryAddress(Long deliveryAddressId) {
+        User user = getAuthenticatedUser(authService.getCurrentUser());
+        DeliveryAddress target = deliveryAddressRepository.findByUserAndIsDefaultTrue(user)
+                .orElseThrow(() -> new DeliveryAddressNotFoundException(ErrorCode.DELIVERY_ADDRESS_NOT_FOUND));
 
-        deliveryAddressRepository.deleteById(deliveryAddress.getId());
+        DeliveryAddress deliveryAddressToDefault = deliveryAddressRepository.findById(deliveryAddressId)
+                .orElseThrow(() -> new DeliveryAddressNotFoundException(ErrorCode.DELIVERY_ADDRESS_NOT_FOUND));
 
-        return true;
+        if (target.hasPermission(user.getId()) && deliveryAddressToDefault.hasPermission(user.getId())) {
+            target.removeDefault();
+            deliveryAddressToDefault.setDefault();
+        } else {
+            throw new UserForbiddenException();
+        }
+    }
+
+    public User getAuthenticatedUser(User user) {
+        if (user == null) {
+            throw new UserUnauthorizedException();
+        }
+        return user;
     }
 }
